% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{Survival Times of Animals}
\format{
This data frame has 48 observations on 3 variables:
\describe{
\item{Poison}{factor levels: I, II, III}
\item{Treatment}{factor levels: A, B, C, D }
\item{Survival}{numeric, animal's survival time measured in 10 hour units}
}
}
\source{
Handbook of Small Datasets; originally in Box and Cox (1964), JRSS A, volume 143, pages 383-430.
}
\usage{
data(animals)
}
\description{
The  data frame records the survival times of 48 animals given different poisons and treatments.
}
\keyword{datasets}
