% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cement.R
\docType{data}
\name{cement}
\alias{cement}
\title{Cement data}
\format{
A dataframe with 21 rows and 2 variables:
\describe{
\item{time}{curing time in days}
\item{strength}{Tensile strength in kilogram per  square centimeter }
}
}
\source{
Hald, A. (1952) Statistical Theory with Engineering Applications, New York: Wiley, 451.
}
\usage{
data(cement)
}
\description{
The data  are from an experiment on the tensile strength of cement.
}
\keyword{datasets}
