% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\docType{data}
\name{engine}
\alias{engine}
\title{Engine data}
\format{
A dataframe with 46 rows and 3 variables:
\describe{
\item{CO}{carbon monoxide}
\item{HC}{hydrocarbons}
\item{NOX}{nitrogen oxides}
}
}
\source{
Prof Jochen Einbeck
}
\usage{
data(engine)
}
\description{
Add description here
}
\keyword{datasets}
