% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scallops.R
\docType{data}
\name{scallops}
\alias{scallops}
\title{Data of Scallops}
\format{
A dataframe with 127 rows and 4 variables:
\describe{
\item{lat}{latitude at which scallops were collected}
\item{long}{longitude at which scallops were collected}
\item{tcatch}{the number of scallops found at a particular location}
\item{y}{the logarithm of the number of scallops found at a particular location}
}
}
\source{
Prof Jochen Einbeck
}
\usage{
data(scallops)
}
\description{
Scallops are small bivalves that live in deep waters and grow in shells, much the way oysters do. The data set consists of 127 locations (represented by two variables, longitude  and latitude) at which scallops were collected in a 1990 survey cruise in the Atlantic continental shelf off Long Island, New York, USA.
}
\keyword{datasets}
