% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2data-package.R
\docType{package}
\name{sm2data-package}
\alias{sm2data}
\alias{sm2data-package}
\title{sm2data: Datasets for Statistical Modelling}
\description{
A collection of datasets for Statistical Modelling practicals run by the
	     Department of Mathematical Sciences at Durham University.
}
\author{
\strong{Maintainer}: Tahani Coolen-Maturi \email{tahani.maturi@durham.ac.uk}

Authors:
\itemize{
  \item Jochen Einbeck \email{jochen.einbeck@durham.ac.uk}
}

}
\keyword{internal}
