% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\docType{data}
\name{student}
\alias{student}
\title{Students in S Statistics 1993-4}
\format{
A dataframe with 43 rows and 8 variables:
\describe{
\item{gender}{numeric	gender coded as 1 or 0}
\item{height}{longitude at which scallops were collected}
\item{stone}{stone}
\item{lb}{lb}
\item{weight}{weight}
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
student
}
\description{
Personal characteristics are recorded for students taking S Statistics in 1993-4. Fitness level was recorded on an informal scale, the details of which now elude me. The important information is that higher numbers are supposed to represent greater levels of fitness.
}
\keyword{datasets}
