% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\docType{data}
\name{engine}
\alias{engine}
\title{Data of engine}
\format{
A dataframe with 127 rows and 4 variables:
\describe{
\item{lat}{latitude at which scallops were collected}
\item{long}{longitude at which scallops were collected}
}
}
\source{
\url{https://www.pnas.org/content/114/7/1512/tab-figures-data}
}
\usage{
engine
}
\description{
Add some text here
}
\keyword{datasets}
